Refactor ForkIdManager
