+++ ethereum/eth/src/main/java/org/hyperledger/besu/ethereum/eth/manager/ForkIdManager.java
@@ -18,15 +18,12 @@ import static com.google.common.base.Preconditions.checkNotNull;
 
 import org.hyperledger.besu.ethereum.chain.Blockchain;
 import org.hyperledger.besu.ethereum.core.Hash;
-import org.hyperledger.besu.ethereum.rlp.BytesValueRLPOutput;
 import org.hyperledger.besu.ethereum.rlp.RLPInput;
-import org.hyperledger.besu.ethereum.rlp.RLPOutput;
+import org.hyperledger.besu.util.EndianUtils;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.function.Function;
 import java.util.function.LongSupplier;
-import java.util.function.Predicate;
 import java.util.stream.Collectors;
 import java.util.zip.CRC32;
 
@@ -39,11 +36,13 @@ public class ForkIdManager {
   private final Hash genesisHash;
   private final List<ForkId> forkAndHashList;
 
-  private final Predicate<ForkId> forkIDChecker;
   private final List<Long> forks;
   private final LongSupplier chainHeadSupplier;
   private long lastHead;
   private ForkId lastComputedForkId;
+  private final long forkNext;
+  private final boolean onlyZerosForkBlocks;
+  private final long highestKnownFork;
 
   public ForkIdManager(final Blockchain blockchain, final List<Long> nonFilteredForks) {
     checkNotNull(blockchain);
@@ -57,13 +56,9 @@ public class ForkIdManager {
             .distinct()
             .sorted()
             .collect(Collectors.toUnmodifiableList());
-    if (onlyZerosForkBlocks(nonFilteredForks)) {
-      this.forkIDChecker = forkId -> true;
-    } else {
-      this.forkIDChecker =
-          createForkIDChecker(
-              blockchain, genesisHash, nonFilteredForks, fs -> this.forks, forkAndHashList);
-    }
+    this.onlyZerosForkBlocks = nonFilteredForks.stream().allMatch(value -> 0L == value);
+    this.forkNext = createForkIds();
+    this.highestKnownFork = !forks.isEmpty() ? forks.get(forks.size() - 1) : 0L;
   }
 
   public ForkId computeForkId() {
@@ -89,33 +84,7 @@ public class ForkIdManager {
   }
 
   @VisibleForTesting
-  ForkId getLatestForkId() {
-    if (forkAndHashList.size() > 0) {
-      return forkAndHashList.get(forkAndHashList.size() - 1);
-    }
-    return null;
-  }
-
-  private static Predicate<ForkId> createForkIDChecker(
-      final Blockchain blockchain,
-      final Hash genesisHash,
-      final List<Long> forks,
-      final Function<List<Long>, List<Long>> sanitizer,
-      final List<ForkId> forkIds) {
-    final List<Long> sanitizedForks = sanitizer.apply(forks);
-    final long forkNext = createForkIds(genesisHash, sanitizedForks, forkIds);
-    return eip2124(blockchain, forkNext, forkIds, highestKnownFork(sanitizedForks));
-  }
-
-  private static boolean onlyZerosForkBlocks(final List<Long> forks) {
-    return forks.stream().allMatch(value -> 0L == value);
-  }
-
-  private static long highestKnownFork(final List<Long> forks) {
-    return !forks.isEmpty() ? forks.get(forks.size() - 1) : 0L;
-  }
-
-  public List<ForkId> getForkAndHashList() {
+  List<ForkId> getForkAndHashList() {
     return this.forkAndHashList;
   }
 
@@ -134,50 +103,33 @@ public class ForkIdManager {
    * @return boolean (peer valid (true) or invalid (false))
    */
   boolean peerCheck(final ForkId forkId) {
-    return forkIDChecker.test(forkId);
+    if (forkId == null || onlyZerosForkBlocks) {
+      return true; // Another method must be used to validate (i.e. genesis hash)
+    }
+    // Run the fork checksum validation rule set:
+    //   1. If local and remote FORK_CSUM matches, connect.
+    //        The two nodes are in the same fork state currently. They might know
+    //        of differing future forks, but that's not relevant until the fork
+    //        triggers (might be postponed, nodes might be updated to match).
+    //   2. If the remote FORK_CSUM is a subset of the local past forks and the
+    //      remote FORK_NEXT matches with the locally following fork block number,
+    //      connect.
+    //        Remote node is currently syncing. It might eventually diverge from
+    //        us, but at this current point in time we don't have enough information.
+    //   3. If the remote FORK_CSUM is a superset of the local past forks and can
+    //      be completed with locally known future forks, connect.
+    //        Local node is currently syncing. It might eventually diverge from
+    //        the remote, but at this current point in time we don't have enough
+    //        information.
+    //   4. Reject in all other cases.
+    if (!isHashKnown(forkId.getHash())) {
+      return false;
+    }
+    return chainHeadSupplier.getAsLong() < forkNext
+        || (isForkKnown(forkId.getNext())
+            && isRemoteAwareOfPresent(forkId.getHash(), forkId.getNext()));
   }
 
-  private static Predicate<ForkId> eip2124(
-      final Blockchain blockchain,
-      final long forkNext,
-      final List<ForkId> forkAndHashList,
-      final long highestKnownFork) {
-    return forkId -> {
-      if (forkId == null) {
-        return true; // Another method must be used to validate (i.e. genesis hash)
-      }
-      // Run the fork checksum validation rule set:
-      //   1. If local and remote FORK_CSUM matches, connect.
-      //        The two nodes are in the same fork state currently. They might know
-      //        of differing future forks, but that's not relevant until the fork
-      //        triggers (might be postponed, nodes might be updated to match).
-      //   2. If the remote FORK_CSUM is a subset of the local past forks and the
-      //      remote FORK_NEXT matches with the locally following fork block number,
-      //      connect.
-      //        Remote node is currently syncing. It might eventually diverge from
-      //        us, but at this current point in time we don't have enough information.
-      //   3. If the remote FORK_CSUM is a superset of the local past forks and can
-      //      be completed with locally known future forks, connect.
-      //        Local node is currently syncing. It might eventually diverge from
-      //        the remote, but at this current point in time we don't have enough
-      //        information.
-      //   4. Reject in all other cases.
-      if (isHashKnown(forkId.getHash(), forkAndHashList)) {
-        if (blockchain.getChainHeadBlockNumber() < forkNext) {
-          return true;
-        } else {
-          if (isForkKnown(forkId.getNext(), highestKnownFork, forkAndHashList)) {
-            return isRemoteAwareOfPresent(
-                forkId.getHash(), forkId.getNext(), highestKnownFork, forkAndHashList);
-          } else {
-            return false;
-          }
-        }
-      } else {
-        return false;
-      }
-    };
-  }
   /**
    * Non EIP-2124 behaviour
    *
@@ -188,21 +140,16 @@ public class ForkIdManager {
     return !peerGenesisHash.equals(genesisHash);
   }
 
-  private static boolean isHashKnown(final Bytes forkHash, final List<ForkId> forkAndHashList) {
+  private boolean isHashKnown(final Bytes forkHash) {
     return forkAndHashList.stream().map(ForkId::getHash).anyMatch(hash -> hash.equals(forkHash));
   }
 
-  private static boolean isForkKnown(
-      final Long nextFork, final long highestKnownFork, final List<ForkId> forkAndHashList) {
+  private boolean isForkKnown(final Long nextFork) {
     return highestKnownFork < nextFork
         || forkAndHashList.stream().map(ForkId::getNext).anyMatch(fork -> fork.equals(nextFork));
   }
 
-  private static boolean isRemoteAwareOfPresent(
-      final Bytes forkHash,
-      final Long nextFork,
-      final long highestKnownFork,
-      final List<ForkId> forkAndHashList) {
+  private boolean isRemoteAwareOfPresent(final Bytes forkHash, final Long nextFork) {
     for (final ForkId j : forkAndHashList) {
       if (forkHash.equals(j.getHash())) {
         if (nextFork.equals(j.getNext())) {
@@ -217,8 +164,7 @@ public class ForkIdManager {
     return false;
   }
 
-  private static long createForkIds(
-      final Hash genesisHash, final List<Long> forks, final List<ForkId> forkIds) {
+  private long createForkIds() {
     final CRC32 crc = new CRC32();
     crc.update(genesisHash.toArray());
     final List<Bytes> forkHashes = new ArrayList<>(List.of(getCurrentCrcHash(crc)));
@@ -230,110 +176,22 @@ public class ForkIdManager {
 
     // This loop is for all the fork hashes that have an associated "next fork"
     for (int i = 0; i < forks.size(); i++) {
-      forkIds.add(new ForkId(forkHashes.get(i), forks.get(i)));
+      forkAndHashList.add(new ForkId(forkHashes.get(i), forks.get(i)));
     }
     long forkNext = 0;
     if (!forks.isEmpty()) {
-      forkNext = forkIds.get(forkIds.size() - 1).getNext();
-      forkIds.add(new ForkId(forkHashes.get(forkHashes.size() - 1), 0));
+      forkNext = forkAndHashList.get(forkAndHashList.size() - 1).getNext();
+      forkAndHashList.add(new ForkId(forkHashes.get(forkHashes.size() - 1), 0));
     }
     return forkNext;
   }
 
   private static void updateCrc(final CRC32 crc, final Long block) {
-    final byte[] byteRepresentationFork = longToBigEndian(block);
+    final byte[] byteRepresentationFork = EndianUtils.longToBigEndian(block);
     crc.update(byteRepresentationFork, 0, byteRepresentationFork.length);
   }
 
   private static Bytes getCurrentCrcHash(final CRC32 crc) {
     return Bytes.ofUnsignedInt(crc.getValue());
-  }
-
-  public static class ForkId {
-    final Bytes hash;
-    final Bytes next;
-    Bytes forkIdRLP;
-
-    private ForkId(final Bytes hash, final Bytes next) {
-      this.hash = hash;
-      this.next = next;
-      createForkIdRLP();
-    }
-
-    public ForkId(final Bytes hash, final long next) {
-      this(hash, Bytes.wrap(longToBigEndian(next)).trimLeadingZeros());
-    }
-
-    public long getNext() {
-      return next.toLong();
-    }
-
-    public Bytes getHash() {
-      return hash;
-    }
-
-    void createForkIdRLP() {
-      final BytesValueRLPOutput out = new BytesValueRLPOutput();
-      writeTo(out);
-      forkIdRLP = out.encoded();
-    }
-
-    public void writeTo(final RLPOutput out) {
-      out.startList();
-      out.writeBytes(hash);
-      out.writeBytes(next);
-      out.endList();
-    }
-
-    public static ForkId readFrom(final RLPInput in) {
-      in.enterList();
-      final Bytes hash = in.readBytes();
-      final long next = in.readLongScalar();
-      in.leaveList();
-      return new ForkId(hash, next);
-    }
-
-    public List<ForkId> asList() {
-      final ArrayList<ForkId> forRLP = new ArrayList<>();
-      forRLP.add(this);
-      return forRLP;
-    }
-
-    @Override
-    public String toString() {
-      return "ForkId(hash=" + this.hash + ", next=" + next.toLong() + ")";
-    }
-
-    @Override
-    public boolean equals(final Object obj) {
-      if (obj instanceof ForkId) {
-        final ForkId other = (ForkId) obj;
-        final long thisNext = next.toLong();
-        return other.getHash().equals(this.hash) && thisNext == other.getNext();
-      }
-      return false;
-    }
-
-    @Override
-    public int hashCode() {
-      return super.hashCode();
-    }
-  }
-
-  // next two methods adopted from:
-  // https://github.com/bcgit/bc-java/blob/master/core/src/main/java/org/bouncycastle/util/Pack.java
-  private static byte[] longToBigEndian(final long n) {
-    final byte[] bs = new byte[8];
-    intToBigEndian((int) (n >>> 32), bs, 0);
-    intToBigEndian((int) (n & 0xffffffffL), bs, 4);
-    return bs;
-  }
-
-  @SuppressWarnings("MethodInputParametersMustBeFinal")
-  private static void intToBigEndian(final int n, final byte[] bs, int off) {
-    bs[off] = (byte) (n >>> 24);
-    bs[++off] = (byte) (n >>> 16);
-    bs[++off] = (byte) (n >>> 8);
-    bs[++off] = (byte) (n);
   }
 }
