+++ util/src/main/java/org/hyperledger/besu/util/EndianUtils.java
@@ -1,0 +1,34 @@
+/*
+ * Copyright ConsenSys AG.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+package org.hyperledger.besu.util;
+
+public class EndianUtils {
+  // next two methods adopted from:
+  // https://github.com/bcgit/bc-java/blob/master/core/src/main/java/org/bouncycastle/util/Pack.java
+  public static byte[] longToBigEndian(final long n) {
+    final byte[] bs = new byte[8];
+    intToBigEndian((int) (n >>> 32), bs, 0);
+    intToBigEndian((int) (n & 0xffffffffL), bs, 4);
+    return bs;
+  }
+
+  @SuppressWarnings("MethodInputParametersMustBeFinal")
+  public static void intToBigEndian(final int n, final byte[] bs, int off) {
+    bs[off] = (byte) (n >>> 24);
+    bs[++off] = (byte) (n >>> 16);
+    bs[++off] = (byte) (n >>> 8);
+    bs[++off] = (byte) (n);
+  }
+}
